requirements.txt
asgiref==3.4.1
certifi==2021.10.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
pika==1.2.0
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
wincertstore==0.2
